requirements.txt
pandas
python-dotenv
mysql-connector-python
sqlalchemy
streamlit
streamlit_option_menu
matplotlib
seaborn
bebel